rename sessions to focus-sessions
