+++ src/modules/focus-sessions/application/services/log-distraction.service.ts
@@ -1,0 +1,41 @@
+import {
+  ForbiddenException,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import {
+  SESSION_REPO_PORT,
+  type ISessionRepository,
+} from '@focus-sessions/domain/ports/session-repo.port';
+import type { Session } from '@focus-sessions/domain/entities/session.entity';
+
+export interface LogDistractionInput {
+  sessionId: string;
+  userId: string;
+  reason: string;
+  estimatedMs: number;
+}
+
+@Injectable()
+export class LogDistractionService {
+  constructor(
+    @Inject(SESSION_REPO_PORT)
+    private readonly sessionRepo: ISessionRepository,
+  ) {}
+
+  async execute(input: LogDistractionInput): Promise<Session> {
+    const session = await this.sessionRepo.findById(input.sessionId);
+    if (!session) throw new NotFoundException('Session not found.');
+    if (session.userId !== input.userId)
+      throw new ForbiddenException('Access denied.');
+
+    const updated = session.addDistraction({
+      reason: input.reason,
+      estimatedMs: input.estimatedMs,
+    });
+
+    await this.sessionRepo.update(updated);
+    return updated;
+  }
+}
