+++ src/modules/focus-sessions/domain/events/session-completed.event.ts
