+++ src/modules/focus-sessions/infrastructure/persistence/session-mongo.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common';
+import { MongooseModule } from '@nestjs/mongoose';
+import {
+  SessionDoc,
+  SessionSchema,
+} from '@focus-sessions/infrastructure/persistence/session.schema';
+
+@Module({
+  imports: [
+    MongooseModule.forFeature([
+      { name: SessionDoc.name, schema: SessionSchema },
+    ]),
+  ],
+  exports: [MongooseModule],
+})
+export class SessionMongoModule {}
