+++ src/modules/goals/infrastructure/workers/goal-workers.module.ts
@@ -8,7 +8,7 @@ import { GOALS_QUEUE_NAME } from '@goals/domain/types/goal-jobs.types';
 import { GOAL_EVENT_PUBLISHER_PORT } from '@goals/domain/ports/goal-event-publisher.port';
 import { GoalMongoModule } from '@goals/infrastructure/persistence/goal-mongo.module';
 import { TaskMongoModule } from '@tasks/infrastructure/persistence/task-mongo.module';
-import { SessionMongoModule } from '@sessions/infrastructure/persistence/session-mongo.module';
+import { SessionMongoModule } from '@focus-sessions/infrastructure/persistence/session-mongo.module';
 
 @Module({
   imports: [
