--- src/modules/sessions/domain/ports/session-event-publisher.port.ts
@@ -1,9 +1,0 @@
-import type { SessionCompletedEvent } from '@sessions/domain/events/session-completed.event';
-
-export const SESSION_EVENT_PUBLISHER_PORT = Symbol(
-  'SESSION_EVENT_PUBLISHER_PORT',
-);
-
-export interface ISessionEventPublisher {
-  publishSessionCompleted(event: SessionCompletedEvent): Promise<void>;
-}
