+++ src/modules/tasks/infrastructure/workers/task-workers.module.ts
@@ -10,7 +10,7 @@ import { TaskMongoModule } from '@tasks/infrastructure/persistence/task-mongo.mo
 // HabitMongoModule registers HabitOccurrence model used by the MARK_HABIT_OCCURRENCE handler.
 import { HabitMongoModule } from '@habits/infrastructure/persistence/habit-mongo.module';
 import { GoalMongoModule } from '@goals/infrastructure/persistence/goal-mongo.module';
-import { SessionMongoModule } from '@sessions/infrastructure/persistence/session-mongo.module';
+import { SessionMongoModule } from '@focus-sessions/infrastructure/persistence/session-mongo.module';
 import { AnalyticsWorkersModule } from '@analytics/infrastructure/workers/analytics-workers.module';
 
 @Module({
